Renombrar namespace de ToastModel, ToastBase y Toast
