+++ SistemaTurnosOnline.Web/Components/ToastComponent/Parent/ToastModel.cs
@@ -1,27 +1,34 @@
 ﻿using Microsoft.AspNetCore.Components;
 using Microsoft.JSInterop;
-using SistemaTurnosOnline.Web.Components.Contracts;
 using SistemaTurnosOnline.Web.Components.Contracts.ToastContracts;
+using SistemaTurnosOnline.Web.Components.Contracts;
 using SistemaTurnosOnline.Web.Extensions;
 
-namespace SistemaTurnosOnline.Web.Components.ToastComponent
+namespace SistemaTurnosOnline.Web.Components.ToastComponent.Parent
 {
-    public abstract class ToastBase : ComponentBase, IHasId, IToastCss, IToastOptions, IShowable
+    public abstract class ToastModel : IHasId, IToastCss, IToastOptions, IShowable
     {
+        public ToastModel(string id, string title, string text)
+        {
+            Id = id;
+            Title = title;
+            Text = text;
+        }
+
         // The ID of the toast element.
-        [Parameter] public string Id { get; set; }
+        public string Id { get; set; }
 
         // The icon to use for the toast.
-        [Parameter] public string Icon { get; set; }
+        public string Icon { get; set; }
 
         // The CSS class to use for the header of the toast.
-        [Parameter] public string HeaderClass { get; set; }
+        public string HeaderClass { get; set; }
 
         // The title to display in the toast.
-        [Parameter] public string Title { get; set; }
+        public string Title { get; set; }
 
         // The text to display in the body of the toast.
-        [Parameter] public string Text { get; set; }
+        public string Text { get; set; }
 
         // Shows the toast using the specified JS runtime.
         public async Task Show(IJSRuntime jsRuntime)
